requirements.txt
asgiref==3.7.2
black==23.10.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.6
django-autoslug==1.9.9
django-environ==0.11.2
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djangorestframework-xml==2.0.0
djoser==2.2.0
idna==3.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
ruff==0.1.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
